Fix EventBased algorithm dropping zero event lumis
The current mask definition performed by the EventBased splitting
algorithm on non-mc input files can lead to missing lumis in the output.
This occurs when an input file has zero events in the last lumis, because
the cmsRun process will stop once it has reached the max events and won't process
those eventless lumis. To fix this, avoid passing the maxEvents parameter
when configuring the mask of the last job in a file, the skipEvents parameter
will suffice and the job will run over the remainder of the file.

Many assertions changed in the unit tests, since we change from having
always a max events to having it only on intermediate jobs in a file.