ReqMgr CustodialSite bug fix.
CustodialSite is an argument not used for long. (Non)CustodialSites have
superseded this, however no one bothered to remove all dead functionality and
parameter definition. On cloned requests, the default value (None) of
CustodialSite request parameter appared and PhEDExInjector was failing on
that. This patch removes CustodialSite from ReqMgr only.