Fix how we delete documents in couchdb
A weird hitch in couch's delete semantics is that if you don't use the
DELETE verb (i.e. you commit with _bulk_docs and set _deleted=true), it
stores the whole data in the delete tombstone. See:

https://issues.apache.org/jira/browse/COUCHDB-1141

Since (AFAIK), we don't ever use _deleted_revs=true to look at things,
it means that we have tons of old data sitting around in couch that's
basically invisible. This patch strips the document before comitting it
so the only thing lefts are _id, _rev and _deleted