Improve ACDC splitting in global queue
Allow two modes of operation dependant on the splitting algorithm
that will be used for the top level task. One is the currently used
mode that splits in fixed sized chunks and the other one is a single
chunk mode that prevents blocking the local queue when splitting
algorithms required wmbs injection before creating jobs.

Also remove the cleanup tasks from merge ACDCs, as it stands now
these won't work and are currently removed manually by ops through
hacky scripts.

Fixes #4043
Fixes #4029 (Not properly but it does fix it)