Bugfix for file information
In some occasions the TaskArchiver fails to delete some file records
correctly from WMBS and some records are left with the default
values in wmbs_file_details. If a second workflow comes with the same
input files it may find zero events files in WMBS which causes
problems in the splitting algo like EventAwareLumiBased. Fix
this by updating the file information in WMBS when trying to add a new file
that is already there. This rarely happens so there is no real cost.