optimize sibgling subscription DAO
Optimize the sibling complete DAO query to not join against the huge wmbs_fileset_files table and only against wmbs_fileset. Also works if the other subscription are cleaned up before (even though this should not happen anymore).

Also change the unit tests to remove the case of multiple SE for a single location, we simple do not support that (never have).