Include performance parameter in workload
First part of #4472, include the performance information that is now already
required but only stored in the schema. The performance information is stored by
task and can be different depending on the type of task.

Tasks like skims and second step TaskChain use their own estimate which is optional
and defaults to the top level value.

This is only the ReqMgr and Workload definition changes. After this we need the changes at the
agent level in the splitters but that is independent of cmsweb schedule.