Fix xrootd option for ACDC
Allow the trust site whitelist as location option to work for ACDC workflows
the ACDC files will have locations according to the site whitelist and blacklist.

Also improve an ACDC view making it compatible to newer versions of couchDB where
documents can't be modified in a view (it was sorting in place a element of the doc
inside the view).

Fixes #4496