ReqMgr request parameter DbsUrl stored in CouchDB
Correction of #4515
e.g. for MonteCarlo workload the DbsUrl was not really stored in CouchDB, even
though it was specified at injection, querying the request was returning None
for DbsUrl. As discussed with @efajardo, DbsUrl argument, once specified,
shall be stored for all types of workflows.

Fixes #4498
Correction of #4380 (DbsUrl should not have been removed, only fixed so that
it's properly stored in CouchDB, as done here now)