Avoid re-running performance killed jobs
Improve the exit code handling in the ErrorHandler,
retrieve a list instead of a single random non-zero exit code.
Compare the list against the exit and pass list.
Change the exit codes for performance kills to make them identifiable,
include them in the default configurable list or exit codes.

Fixes #4473