Make robust merge jobs
Add the skipBadFiles flag to the input source in merge jobs
in order to avoid failing the merge for a single missing file.
When a successful merge job comes back with skipped files
these are added to the ACDC collection for the merge task
for posterior recovery.

Additionally, now CMSSW steps will be declared failed
if no output files are produced. This way the scenario
where all files were skipped is not counted as a success and also
allows to catch glitches when the cmsRun process fails to process the input
but stil returns with exit code 0.

This is only enabled for ParentlessMergeBySize.

Fixes #4476