Store correctly the SizePerEvent and friends
Make sure that SizePerEvent is stored correctly in CouchDB and Oracle
if present, it goes in hand with RequestSizeFiles and RequestNumEvents
but these two are not really present in any request schema
so they will be always zero, only SizePerEvent is mandatory.

Also fix the clone and assign of requests pre-upgrade.