Fix pileup dataset location tracking
On workflows where the first task is production, the
dbsUrl is not available in the top level task. Therefore
the WorkQueue was choking when trying to get the locations
since it couldn't find a dbsUrl.

Change it to properly use the dbsUrl that is shipped in the stephelper
information with the pileup datasets.