Fix sizePerEvent for MC
MC requests may have very low filter efficiency
so the estimates for disk space will be too high
since the current calculation assumes the definition of:

"Size per event is the size of one event in the output"

However, a more proper definition would be:

- "Size per event is the contribution to the output size for every event in the
   input, if there is no input (i.e. production) then it is the size of every
   event in the output"

Under this definition we must adjust the sizePerEvent according to the efficiency
value in MonteCarlo/LHEStepZero but for all the other processing requests
we expect size per event values according to the above definition (e.g. in
the skims, the skim efficiency should be considered in the sizeperEvent).