Fix location update for Pileup on Generation
When the top level task is a Production task with no DBS Url,
the location updater crashes since there is no DBS to do the
location update. Fix that by including a DBS url in the workqueue
element even if not present in the top level task, it will use one from
the pileup datasets.