Use stale views in the TaskArchiver
In agents with high load, the JobDump and FWJRDump design views
are always updating and therefore stale views must be used there.
The TaskArchiver data collection happens late enough that a stale
view shouldn't cause problems with missing data.

Unfortunately, CouchDB doesn't accept stale on lists. For those
we have to catch the timeout exceptions and move on.