Workload validation refactoring
Implement a new way of defining and validitng the request arguments
in the WMSpec objects, reorganize them to make it simpler
and clearer. It simplifies some parts of the ReqMgr1 when
creating requests and reduces the meddling with request parameters.

Basically what a requestor gives can be passed to the specs
and these will take care of parsing and validation procedures.

Major changes are:

- LHEStepZero is removed and contained in MonteCarlo
- Splitting arguments in creation have their names simplified
- EnableDQMHarvest becomes EnableHarvesting and needs a True or False value

Fixes #4224