make TasArchiver Reco performance upload optional
Make the upload of Reco performance data to the Dashboard from the TaskArchiver optional.
There is currently a problem with either the URL or the code, but in any case this needs
to be optional because I don't want to have it active in (every) Tier0 replays. So if
the url is not overidden, do not use a default but simply do not upload.