Revise the FNAL plugin for better protocol enforcement
We now set the stageout method in the following way:

 To xrdcp only if the PFN starts with root://
 To srm   only if the PFN starts with srm://
 To local for all others.

We also forbid (and return an error) when the source method is srm and
the target is xrdcp (and vice-versa). Other combinations are allowed.

This commit also removes some obsolete code, adds a little better
debugging, and moves a code path earlier if it's being invoked for staging in.