ignore location for job submission and only use sitewhitelist and siteblacklist
Add a new parameters TrustSitelists in StdBase that is propogated to the process task jobsplitter (just like the sitewhitelist and siteblacklist). From there it makes it into the job object and is picked up by the JobSubmitter, where it forces use of the site lists without taking data location into account.

Also remove the site lists from merge jobs and make applying the site lists to processing tasks optional (later tasks in a task chain might not want to use them).