cleanup of the use of rowcount
The way we handled rowcount for the results of a query was somewhat hacky and didn't work with
newer versions of SqlAlchemy and Oracle. As rowcount was only used in one place in the code and
it wasn't even needed there, I completely removed the rowcount field and the hacky code.