some preliminary work for a copy-delete type subscription
Normally there are only two subscription modes supported, copy or move.
We want to implement a delayed move, consisting of a copy first, followed by
a delete from the source site once the workflow creating the data is done.

These are some prelimary changes for this. The main changes are in the schema,
adding a delete_blocks column to dbsbuffer_dataset_subscription and a
deleted column to dbsbuffer_block. To make this work, we also needed to
add a FK constrained dataset_id column to dbsbuffer_block, which then
forced us to make some code changes to properly fill this.

These code changes then triggered some neccessary cleanup work. Tried to
keep it as minimal as possible though for now, more is to come.

Also fix a few unit tests to work with the changes.