use CMSSW python 2.7 for runtime
Newer CMSSW releases include python 2.7 as external. Some of their configuration and especially code in Config.DataProcessing starts depending on this newer python version and will not work anymore with the default python 2.6 that comes with the DMWM externals. Change the WMCore runtime code to use the CMSSW provided python 2.7 (if it exists) to run the parts of the runtime code that is close to CMSSW.