implement a copy-delete type subscription mode
Normally there are only two subscription modes supported, copy or move. This implements a delayed move, consisting of a copy first, followed by a delete from the injection site once the workflow creating the data is done and the block is fully transferred to all sites the WMAgent made a subscription for.

If there is an active (outside the agent) subscription to the injection site, the delete is skipped.