Remove sitesFromStorageElements
As we wish to deal only with PSNs and PNNs we no longer need this function and it can be replaces with a query to the SiteDB PNN to PSN mapping.

This partially fixes https://github.com/dmwm/WMCore/issues/6350

Convert file block locations to psn

as part of https://github.com/dmwm/WMCore/issues/6350 must also convert the file block locations

Convert listFileBlockLocations to PSN in Dataset

This is also needed as part of https://github.com/dmwm/WMCore/issues/6350
A white list check is done so must compare with the PSN of the data location.

Handle legacy se names from dbs

before dbs is fully switched to using PNNs we still need to handle the case when it gives us se names.
This addresses https://github.com/dmwm/WMCore/issues/6349

fix location mapping