First draft at tracking blow-up factor.
We take blow-up factor to be the number of additional jobs in a
task chain after the jobs in the first task.  The latter steps in
the task chain should affect what is acquired for work now - esp
as the blow-up factor may be >100 in the cases of LHE.

This approach is not foolproof, I think.  WQE's already in WMBS
may only count the jobs currently in WMBS still.  If nothing else,
this will slow the acquisition of such workflows to the point where
they are load-balanced across agents.  Not clear that it would
prevent an agent from taking on too much work if it's a single-agent
team.