Allow runtime environment to resize the cores in the job.
This allows WMRuntime to resize the number of threads in the pset
*if* the job is in the condor runtime environment (_CONDOR_MACHINE_AD
env var is set and the slot's ad is in the referenced file), the
Cpus attribute is set to an integer, and WMCore_ResizeJob is set to
'true'.

So, by default this does nothing.  However, Ops can then explicitly
enable job resizing (and setting RequestCpus and RequestMemory
appropriately to reflect this) for a given workflow or subtask in
condor.  Since I suspect we'll need to experiment a bit with the
correct values for RequestMemory / RequestCpus, I don't currently
change those in the CondorPlugin.

This just gives us the flexibility to explore the phase space.

Cases of resizeCores tested by hand:
  - When _CONDOR_MACHINE_AD env var is missing.
  - When machine ad is unparseable.
  - When resize job attribute is missing, true, or false.
  - Non-parseable values for Cpus or WMCore_ResizeJob
  - Cpus is greater than, equal to, or less than numCores