Change PyCondorPlugin to be able to dynamically resize jobs.
This changes Cpus, Memory, and MaxWallTimeMins to be appropriate
for any number of cores that might be available at matchmaking time.

For these to be effective, one will need to set WMCore_ResizeJob
to be True in the ad.  This defaults to False right now and would
have to be enabled by hand.  A separate update will allow this to
be specified within the workflow.

Running / completed jobs will have RequestCpus evaluate to the number
of cores on the machine where the job was matched.

This means that monitoring and accounting systems can continue to look solely
at this attribute and have it "make sense" for running jobs instead of
needing to pick out match expression attributes.

We factor out original memory and walltime estimates into separate attrs;
this allows us to keep the raw numbers in a separate classad; HTCondor
can then do rounding and avoid breaking auto-clustering.