implement repack output straight to error dataset
Repack jobs stageout directly to merged for files larger than 2GB. If these files are larger than 12GB we do not want them to be included in the regular RAW output dataset and we also don't want them to be prompt reconstructed. Therefore implement an override in the JobAccountant that changes the DBS Primary Dataset for such files and also skips adding them to the filesets used for PromptReco.