Switch from UUID1 to UUID4.
The UUID1 algorithm generates UUIDs based on a host's MAC address
and a timestamp.  While physical MACs are all indeed unique, Docker
containers very commonly use the same MAC address.  In practice, given
the number of sites converting to Docker, this removes a lot of the
uniqueness and has led to observed collisions.

This changes to UUID4, where the UUID is generated entirely from
the RNG.