[py2py3] Utils - compatible with both python2 and python3
This commit is related to #9731

The following files have been automatically modified with futurize.

	modified:   src/python/Utils/CertTools.py
	modified:   src/python/Utils/EmailAlert.py
	modified:   src/python/Utils/ExtendedUnitTestCase.py
	modified:   src/python/Utils/IteratorTools.py
	modified:   src/python/Utils/MathUtils.py
	modified:   src/python/Utils/MemoryCache.py
	modified:   src/python/Utils/Pipeline.py
	modified:   src/python/Utils/ProcessStats.py
	modified:   src/python/Utils/Signals.py
	modified:   src/python/Utils/Throttled.py
	modified:   src/python/Utils/Timers.py
	modified:   src/python/Utils/Utilities.py

After applying futurize, the following manual changes
have been applied.

1. Dictionaries

replaced mydict.$operator() with view$operator(mydict)

This is a temporary change and will not be needed anymore
after dropping python2 support. This is used for performance
reasons when running python2 with idioms that are compatibles
with python3.

Resources:
* http://python-future.org/compatible_idioms.html#dictionaries
* http://python-future.org/what_else.html#dict

2. Int and Long

In python3 they have been merged. Int type in py3 is the same as
long type in py2

Resources:
* http://python-future.org/compatible_idioms.html#long-integers
* http://python-future.org/what_else.html#int

3. Bytes and Unicode strings

`from builtins import str` dramatically changed the meaning of
the return of `convertFromUnicodeToStr` function in
src/python/Utils/IteratorTools.py

This function is not used at the moment in dmwm/WMCore,
but it has been updated to use python3-compatible definition
of bytes and unicode strings.

4. Update some older idioms

We use jenkins to check if in the changed files there are some
idioms that are from pre-python 2.6 era. It suggested the change

- if type(expected_obj) != type(actual_obj):
+ if not isinstance(expected_obj, type(actual_obj)):

which comes from 2to3

Resource:
* https://docs.python.org/3/library/2to3.html#2to3fixer-idioms