[py2py3] update stdlib - not pickle - test
WMCore does not depend on py2 stdlib anymore, apart from pickle.

This is related to the PR on github dmwm/WMCore #9762

The following stages of futurize have been applied:

* libfuturize.fixes.fix_future_standard_library
* libfuturize.fixes.fix_future_standard_library_urllib

Manual changes have been done to revert changes affecting pickle.

It would have been impossible to completely decouple the two fixers,
since if we run `fix_future_standard_library` on a file containing httplib
(that should be converted to http.client) (this would add install_aliases())
and the same module also contains `import urlparse`, then `urlparse` should
be translated to `urllib.parse`. This second change is done by the second fixer.

Hence, fix_future_standard_library_urllib fixers has been applied in the same
PR as fix_future_standard_library.

Backward compatibility: look at the PR section "is it backward compatible"
for further details.
TL;DR this should be safe if you do not check the exact type of exception
that is raised by UnpackUserTarball

Notes

* urllib: we opted for using in python2 the backproted version of py3 urllib,
  provided by futurize.backports

Details

1. WMConfigCache_t

When we want to read a file available at a local path, it is
necessary to prepend the path/url with the prefix

* "file:./" if it is a relative path
* "file://" if it is an absolute path

This change is now implemented in unit tests only.
It will be extended throughout the WMCore codebase in a following commit

2. REST_t mathces new urllib interface

Before changes:
* test REST api with 'list?int=a' and urllib2.
  - HTTPError exception catched by self.assertRaises.
  - No introspection
* test REST api with 'list1?int=a' and urllib
  - no exception raised
  - introspection of the response with self.assertEqual

After futurize: replace py2 version of urllib and urllib2 with
python3 urllib library, which behaves as urllib2 in py2,
which means that in case of an http error, a HTTPError
exception is raised.
This means that we need to manually change the code that
takes care of checking if a request returns an error.

We decided to keep both 'list?int=a' and 'list1?int=a',
which raise the same exception, but provide a different messages.
We keep both self.assertRaises and exception introspection.